check: move verification of individual pack file to repository
